+++ baseline/modules.py
@@ -1,0 +1,94 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Mon Jun 11 12:29:33 2018
+
+@author: akash
+"""
+
+import math
+
+import torch
+import torch.nn as nn
+import torch.nn.functional as F
+
+from functions import *
+
+
+class BinaryTanh(nn.Module):
+    def __init__(self):
+        super(BinaryTanh, self).__init__()
+        self.hardtanh = nn.Hardtanh()
+
+    def forward(self, input):
+        output = self.hardtanh(input)
+        output = binarize(output)
+        return output
+        
+
+class BinaryLinear(nn.Linear):
+
+    def forward(self, input):
+        binary_weight = binarize(self.weight)
+        if self.bias is None:
+            return F.linear(input, binary_weight)
+        else:
+            return F.linear(input, binary_weight, self.bias)
+
+    def reset_parameters(self):
+        # Glorot initialization
+        in_features, out_features = self.weight.size()
+        stdv = math.sqrt(1.5 / (in_features + out_features))
+        self.weight.data.uniform_(-stdv, stdv)
+        if self.bias is not None:
+            self.bias.data.zero_()
+
+        self.weight.lr_scale = 1. / stdv
+
+
+class BinaryConv2d(nn.Conv2d):
+
+    def forward(self, input):
+        bw = binarize(self.weight)
+        return F.conv2d(input, bw, self.bias, self.stride,
+                               self.padding, self.dilation, self.groups)
+
+    def reset_parameters(self):
+        # Glorot initialization
+        in_features = self.in_channels
+        out_features = self.out_channels
+        for k in self.kernel_size:
+            in_features *= k
+            out_features *= k
+        stdv = math.sqrt(1.5 / (in_features + out_features))
+        self.weight.data.uniform_(-stdv, stdv)
+        if self.bias is not None:
+            self.bias.data.zero_()
+
+        self.weight.lr_scale = 1. / stdv
+
+
+#MODEL
+# class Model(nn.Module):
+#     def __init__(self, num_labels, dim_features):
+#         super(Model, self).__init__()
+#         self.dropout = nn.Dropout(0.3)
+#         self.fc = BinaryLinear(dim_features, num_labels)
+        
+#     def forward(self, x):
+#         out=self.dropout(x)
+#         out = self.fc(out)
+#         return out
+       
+
+class Model(nn.Module):
+
+    def __init__(self, num_labels, dim_features, neurons_hl):
+        super(Model, self).__init__()
+        print(dim_features, neurons_hl)
+        self.fc1 = BinaryLinear(dim_features, neurons_hl)
+        self.fc2 = BinaryLinear(neurons_hl, num_labels)
+
+    def forward(self, x):
+        x = F.relu(self.fc1(x))
+        x = self.fc2(x)
+        return x
